apps/api/views.py
from django.http import JsonResponse
from django.views.decorators.csrf import csrf_exempt
from google.cloud import aiplatform
from vonage import Client, Voice

@csrf_exempt
def ai_diagnosis(request):
    # Placeholder for Google Gemini API integration
    # You'll need to implement the actual API call here
    diagnosis = "Sample AI diagnosis"
    return JsonResponse({'diagnosis': diagnosis})

@csrf_exempt
def create_video_session(request):
    # Placeholder for Vonage API integration
    # You'll need to implement the actual API call here
    client = Client(key=settings.VONAGE_API_KEY, secret=settings.VONAGE_API_SECRET)
    voice = Voice(client)
    # Create a session and return session details
    session_id = "sample_session_id"
    return JsonResponse({'session_id': session_id})