requirements.txt
annotated-types==0.7.0
asgiref==3.7.2
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==4.0.0
cryptography==42.0.8
dataclasses==0.6
Deprecated==1.2.14
deprecation==2.1.0
Django==5.0.2
docstring_parser==0.16
google-ai-generativelanguage==0.4.0
google-api-core==2.19.1
google-auth==2.30.0
google-cloud-aiplatform==1.57.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-generativeai==0.3.1
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gotrue==0.2.0
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
iniconfig==2.0.0
numpy==2.0.0
packaging==24.1
pluggy==1.5.0
postgrest-py==0.4.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
pytest==6.2.5
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.1
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
rsa==4.9
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
supabase-py==0.0.2
toml==0.10.2
tqdm==4.66.4
typing_extensions==4.9.0
urllib3==1.26.19
vonage==2.5.5
websockets==9.1
wrapt==1.16.0
